+++ example/lib/main.dart
@@ -30,12 +30,25 @@ class TestPage extends StatelessWidget {
       ),
       body: Center(
         child: Container(
-          width: 400,
+          width: 600,
           child: Center(
             child: SwipeDeck(
               startIndex: 3,
-              aspectRatio: 4 / 3,
-              emptyIndicator: Container(child: Center(child: Text("Nothing Here"),),),
+              emptyIndicator: Container(
+                child: Center(
+                  child: Text("Nothing Here"),
+                ),
+              ),
+              cardSpreadInDegrees: 5, // Change the Spread of Background Cards
+              onSwipeLeft: (){
+                print("USER SWIPED LEFT -> GOING TO NEXT WIDGET");
+              },
+              onSwipeRight: (){
+                print("USER SWIPED RIGHT -> GOING TO PREVIOUS WIDGET");
+              },
+              onChange: (index){
+                print(IMAGES[index]);
+              },
               widgets: IMAGES
                   .map((e) => GestureDetector(
                         onTap: () {