+++ lib/constants.dart
@@ -1,0 +1,2 @@
+const PI = 3.14;
+const DEFAULT_SPREAD = 0.0872665;