+++ lib/data_holder.dart
@@ -1,8 +1,14 @@
 import 'package:flutter/foundation.dart';
+import 'constants.dart';
 
 class TransformData extends ChangeNotifier {
   bool _leftDrag = false;
   double _tDelta = 0;
+  double spreadRadians = DEFAULT_SPREAD; // 5 Degrees
+
+  TransformData({required double spreadRadians}) {
+    this.spreadRadians = spreadRadians;
+  }
 
   get isLeftDrag => this._leftDrag;