+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - October 22, 2021
+
+* onSwipeLeft, onSwipeRight Callbacks
+* onChange Callback that provides you with the Current Index
+* cardSpreadInDegrees - to change the Spread Radius of Background Widgets [Min (2) , Max(10)]
+
 ## [1.0.0] - September 25, 2021
 
 * Initial Release